Bashed it out
Had to alter the original shell version not to include it into the prompt, but just echo the result. Also removed the `sort -R` in favor of _real_ shuffling of the lines using `shuf`

Also removed the hardcoded raffle.txt since its now injected as a paramter.
Updated the docs to reflect...

Since this changes the original code alot... one might say this is a new bash rafffler...  your call i guess ;)